rename packages to be plural ("interface" package name causes compilation errors)
