+++ app/src/main/java/com/pickledgames/stardewvalleyguide/activities/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.pickledgames.stardewvalleyguide.activity
+package com.pickledgames.stardewvalleyguide.activities
 
 import android.os.Bundle
 import android.support.design.widget.BottomNavigationView
@@ -6,7 +6,7 @@ import android.support.v4.app.Fragment
 import android.support.v7.app.AppCompatActivity
 import com.ncapdevi.fragnav.FragNavController
 import com.pickledgames.stardewvalleyguide.R
-import com.pickledgames.stardewvalleyguide.fragment.FriendshipsFragment
+import com.pickledgames.stardewvalleyguide.fragments.FriendshipsFragment
 import kotlinx.android.synthetic.main.activity_main.*
 
 class MainActivity : AppCompatActivity() {
