+++ app/src/main/java/com/pickledgames/stardewvalleyguide/fragments/FriendshipsFragment.kt
@@ -1,4 +1,4 @@
-package com.pickledgames.stardewvalleyguide.fragment
+package com.pickledgames.stardewvalleyguide.fragments
 
 import android.os.Bundle
 import android.support.v4.app.Fragment
@@ -6,7 +6,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.pickledgames.stardewvalleyguide.R
-import com.pickledgames.stardewvalleyguide.adapter.FriendshipsAdapter
+import com.pickledgames.stardewvalleyguide.adapters.FriendshipsAdapter
 import kotlinx.android.synthetic.main.fragment_friendships.*
 
 class FriendshipsFragment : Fragment() {
