+++ app/src/main/java/com/pickledgames/stardewvalleyguide/interfaces/Sortable.kt
@@ -1,0 +1,5 @@
+package com.pickledgames.stardewvalleyguide.interfaces
+
+interface Sortable {
+    fun sort()
+}
