--- app/src/main/java/com/pickledgames/stardewvalleyguide/model/Birthday.kt
@@ -1,8 +1,0 @@
-package com.pickledgames.stardewvalleyguide.model
-
-import com.pickledgames.stardewvalleyguide.enum.Season
-
-data class Birthday(
-        val season: Season,
-        val day: Int
-)
