+++ app/src/main/java/com/pickledgames/stardewvalleyguide/models/Birthday.kt
@@ -1,0 +1,8 @@
+package com.pickledgames.stardewvalleyguide.models
+
+import com.pickledgames.stardewvalleyguide.enums.Season
+
+data class Birthday(
+        val season: Season,
+        val day: Int
+)
